ml-anomaly-detection/services/causal_inference_service.py
#!/usr/bin/env python3
"""
FIDPS Causal Inference Service
Integrates Causal Inference Engine with ML service for real-time RCA
"""

import json
import logging
from typing import Dict, List, Optional, Any
from datetime import datetime
from kafka import KafkaProducer, KafkaConsumer
import pandas as pd

from models.causal_inference_engine import (
    CausalInferenceEngine,
    RootCauseAnalysis,
    MitigationRecommendation
)

logger = logging.getLogger(__name__)


class CausalInferenceService:
    """Service for real-time causal inference and root cause analysis"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Initialize CIE
        self.cie = CausalInferenceEngine(config.get('cie', {}))
        
        # Kafka connections
        self.producer: Optional[KafkaProducer] = None
        self.consumer: Optional[KafkaConsumer] = None
        
        # Database connection (for historical data)
        self.db_connection = None
        
        # Metrics
        self.metrics = {
            'rca_analyses': 0,
            'recommendations_generated': 0,
            'errors': 0
        }
    
    def initialize(self):
        """Initialize service connections"""
        try:
            # Kafka producer
            self.producer = KafkaProducer(
                bootstrap_servers=self.config['kafka']['bootstrap_servers'],
                value_serializer=lambda x: json.dumps(x).encode('utf-8')
            )
            
            # Kafka consumer (for anomaly events)
            self.consumer = KafkaConsumer(
                'ml-anomalies',
                'damage-predictions',
                bootstrap_servers=self.config['kafka']['bootstrap_servers'],
                value_deserializer=lambda x: json.loads(x.decode('utf-8')),
                group_id='causal-inference-service',
                auto_offset_reset='latest'
            )
            
            self.logger.info("Causal Inference Service initialized")
            
        except Exception as e:
            self.logger.error(f"Error initializing service: {e}")
            raise
    
    def process_anomaly(self, anomaly_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Process anomaly event and perform RCA
        
        Args:
            anomaly_data: Anomaly event data
        
        Returns:
            Dictionary containing RCA results and recommendations
        """
        try:
            self.logger.info(f"Processing anomaly for RCA: {anomaly_data.get('anomaly_id')}")
            
            # Get historical context
            historical_data = self._get_historical_context(
                anomaly_data.get('well_id'),
                anomaly_data.get('timestamp')
            )
            
            # Perform RCA
            damage_type = anomaly_data.get('damage_type', 'DT-02')
            rca = self.cie.analyze_root_cause(
                anomaly_data=anomaly_data,
                historical_context=historical_data,
                damage_type=damage_type
            )
            
            # Generate mitigation recommendations
            recommendations = self.cie.generate_mitigation_recommendations(rca)
            
            # Create result
            result = {
                'rca': {
                    'anomaly_id': rca.anomaly_id,
                    'well_id': rca.well_id,
                    'timestamp': rca.timestamp.isoformat(),
                    'primary_root_cause': rca.primary_root_cause,
                    'root_cause_type': rca.root_cause_type.value,
                    'root_cause_confidence': rca.root_cause_confidence,
                    'contributing_factors': rca.contributing_factors,
                    'evidence_summary': rca.evidence_summary,
                    'statistical_significance': rca.statistical_significance,
                    'correlation_analysis': rca.correlation_analysis
                },
                'mitigation_recommendations': [
                    {
                        'recommendation_id': rec.recommendation_id,
                        'action': rec.action,
                        'description': rec.description,
                        'priority': rec.priority.value,
                        'expected_effectiveness': rec.expected_effectiveness,
                        'expected_risk_reduction': rec.expected_risk_reduction,
                        'required_parameters': rec.required_parameters,
                        'estimated_implementation_time': rec.estimated_implementation_time,
                        'confidence': rec.confidence
                    }
                    for rec in recommendations
                ],
                'processing_timestamp': datetime.now().isoformat()
            }
            
            # Publish to Kafka
            self._publish_rca_results(result)
            
            # Update metrics
            self.metrics['rca_analyses'] += 1
            self.metrics['recommendations_generated'] += len(recommendations)
            
            return result
            
        except Exception as e:
            self.logger.error(f"Error processing anomaly for RCA: {e}")
            self.metrics['errors'] += 1
            raise
    
    def _get_historical_context(
        self,
        well_id: str,
        timestamp: Optional[datetime] = None
    ) -> Optional[pd.DataFrame]:
        """Get historical data for causal analysis"""
        # TODO: Implement database query to get historical data
        # For now, return None (will use knowledge base only)
        return None
    
    def _publish_rca_results(self, results: Dict[str, Any]):
        """Publish RCA results to Kafka topics"""
        try:
            # Publish to RCA topic
            self.producer.send('rca-results', value=results)
            
            # Publish mitigation recommendations separately
            for rec in results['mitigation_recommendations']:
                self.producer.send('mitigation-recommendations', value=rec)
            
            self.producer.flush()
            
        except Exception as e:
            self.logger.error(f"Error publishing RCA results: {e}")
    
    def start(self):
        """Start the service (process messages from Kafka)"""
        self.logger.info("Starting Causal Inference Service...")
        
        if not self.consumer:
            self.initialize()
        
        try:
            for message in self.consumer:
                try:
                    anomaly_data = message.value
                    result = self.process_anomaly(anomaly_data)
                    self.logger.info(f"RCA completed: {result['rca']['primary_root_cause']}")
                except Exception as e:
                    self.logger.error(f"Error processing message: {e}")
        except KeyboardInterrupt:
            self.logger.info("Stopping Causal Inference Service...")
        finally:
            self.stop()
    
    def stop(self):
        """Stop the service"""
        if self.producer:
            self.producer.close()
        if self.consumer:
            self.consumer.close()

