ml-anomaly-detection/requirements.txt
# FIDPS ML Anomaly Detection Service Dependencies

# Core ML Libraries
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4
joblib==1.3.2

# Deep Learning Support (optional)
torch==2.1.2
torchvision==0.16.2

# Time Series Analysis
statsmodels==0.14.1
tslearn==0.6.2

# Kafka Integration
kafka-python==2.0.2

# Database Connectivity
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# Configuration Management
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.7

# HTTP and API
requests==2.31.0
fastapi==0.108.0
uvicorn==0.25.0

# Data Validation
jsonschema==4.20.0

# Serialization
cloudpickle==3.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development Tools
black==23.12.1
flake8==7.0.0

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Security
cryptography==41.0.8

# Utilities
tqdm==4.66.1
click==8.1.7

# Anomaly Detection Specialized
pyod==1.1.3

# MLflow for MLOps
mlflow>=2.8.0

# Causal Inference (optional - DoWhy)
# dowhy>=0.11.0  # Uncomment if needed

